:recycle: rename to menstruation_edit
